Move method from defined class to the method receiver
This leads to methods defined on a superclass (or included module) to be
defined at the child class if the method call was send to this class and
if the class does not implement the method itself.

This might be bad if developing on the module itself because we do not
know the parameter types anymore. Maybe in the future we want to define
it at both modules or model the inheritance chain correctly.