Collapse all booleans to bool type in RBS
It doesn't really make sense to use literal true/false types because if
those values would always be true/false, why use a boolean?

Also this clashes with existing types. Say in our traces a method always
returns true, but the last statement is `1.zero?`. The "zero?" method is
typed as bool but our method would expect true as return type which
leads to a type error.