Do not add members if they are already defined
For example we do not need to add `include Enumerable` to the definition
of `Array`.

Normally those unnecessary members do not cause problems but if they are
generic and already correctly use generic types we destroy this by
always assuming untyped type arguments. In those cases (for example with
Enumerable) it makes sense to not include those members as the original
signature is probably better.